add About page placeholder and link
